+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - (**Server Update**) Add an option to disable showing information when the server got updated
 - (**Manga**) Add an option to include client data during migration
 - (**Migrate**) Add bulk migration
+- (**Reader**) Respect the safe inset area (e.g., top-notch) of devices
 
 ### Changed